+++ taskflowbackend/tests/test_tasks.py
@@ -1818,23 +1818,11 @@ class TestBatchValidateChecksumsTask(
         self.assertEqual(result, True)
         self.zone.refresh_from_db()
         self.assertEqual(
-            self.zone.status_info, DEFAULT_STATUS_INFO[ZONE_STATUS_ACTIVE]
+            self.zone.status_info,
+            DEFAULT_STATUS_INFO[ZONE_STATUS_ACTIVE] + ' (1/1: 100%)',
         )
 
     # TODO: Test with SHA256 checksum (see #2170)
-
-    @override_settings(TASKFLOW_ZONE_PROGRESS_INTERVAL=0)
-    def test_validate_progress(self):
-        """Test validating checksums with progress indicator"""
-        self.make_checksum_object(self.obj)
-        self.add_task(**self.task_kw)
-        result = self.run_flow()
-        self.assertEqual(result, True)
-        self.zone.refresh_from_db()
-        self.assertEqual(
-            self.zone.status_info,
-            DEFAULT_STATUS_INFO[ZONE_STATUS_ACTIVE] + ' (1/1)',
-        )
 
     def test_validate_invalid_in_file(self):
         """Test validating checksums with invalid checksum in file"""
@@ -2600,6 +2588,9 @@ class TestBatchMoveDataObjectsTask(
     @override_settings(TASKFLOW_ZONE_PROGRESS_INTERVAL=0)
     def test_execute_progress(self):
         """Test moving with progress indicator"""
+        # Create checksum objects
+        chk_obj = self.make_checksum_object(self.batch_obj)
+        chk_obj2 = self.make_checksum_object(self.batch_obj2)
         self.assertEqual(
             self.zone.status_info, DEFAULT_STATUS_INFO[ZONE_STATUS_ACTIVE]
         )
@@ -2610,7 +2601,12 @@ class TestBatchMoveDataObjectsTask(
                 'landing_zone': self.zone,
                 'src_root': self.batch_src_path,
                 'dest_root': self.batch_dest_path,
-                'src_paths': [self.batch_obj_path, self.batch_obj2_path],
+                'src_paths': [
+                    self.batch_obj_path,
+                    chk_obj.path,
+                    self.batch_obj2_path,
+                    chk_obj2.path,
+                ],
                 'access_name': IRODS_ACCESS_READ_IN,
                 'user_name': DEFAULT_USER_GROUP,
                 'irods_backend': self.irods_backend,
@@ -2622,8 +2618,8 @@ class TestBatchMoveDataObjectsTask(
         self.zone.refresh_from_db()
         self.assertEqual(
             self.zone.status_info,
-            DEFAULT_STATUS_INFO[ZONE_STATUS_ACTIVE] + ' (2/2)',
-        )
+            DEFAULT_STATUS_INFO[ZONE_STATUS_ACTIVE] + ' (2/2: 100%)',
+        )  # Checksum files should not be counted
 
 
 class TestBatchCalculateChecksumTask(
@@ -2672,9 +2668,8 @@ class TestBatchCalculateChecksumTask(
         self.assertIsNotNone(obj.replicas[0].checksum)
         self.assertEqual(obj.replicas[0].checksum, self.get_checksum(obj))
         self.zone.refresh_from_db()
-        self.assertEqual(
-            self.zone.status_info,
-            DEFAULT_STATUS_INFO[ZONE_STATUS_ACTIVE] + ' (1/1)',
+        self.assertIn(
+            DEFAULT_STATUS_INFO[ZONE_STATUS_ACTIVE], self.zone.status_info
         )
 
     def test_calculate_twice(self):
@@ -2726,7 +2721,7 @@ class TestBatchCalculateChecksumTask(
         self.zone.refresh_from_db()
         self.assertEqual(
             self.zone.status_info,
-            DEFAULT_STATUS_INFO[ZONE_STATUS_ACTIVE] + ' (1/1)',
+            DEFAULT_STATUS_INFO[ZONE_STATUS_ACTIVE] + ' (1/1: 100%)',
         )